Add tests for EVM portable signatures
To run the tests it is necessary to copy the kernel private key from
certs/signing_key.pem to the tests/ directory.

Only the tests with a compatible EVM mode will be executed, unless
the EVM_CHANGE_MODE variable is set to 1. In this case, the script attempts
to change the current EVM mode.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>